+++ CHANGELOG.md
@@ -11,7 +11,7 @@ New features:
 -
 -
 -
--
+- Added support for identity impersonation when using workload identity federation (snowflakedb/gosnowflake#1652).
 -
 
 Bug fixes: